+++ core/types/appstate.go
@@ -27,14 +27,13 @@ type Validator struct {
 }
 
 type Candidate struct {
-	RewardAddress  Address  `json:"reward_address"`
-	OwnerAddress   Address  `json:"owner_address"`
-	TotalBipStake  *big.Int `json:"total_bip_stake"`
-	PubKey         Pubkey   `json:"pub_key"`
-	Commission     uint     `json:"commission"`
-	Stakes         []Stake  `json:"stakes"`
-	CreatedAtBlock uint     `json:"created_at_block"`
-	Status         byte     `json:"status"`
+	RewardAddress Address  `json:"reward_address"`
+	OwnerAddress  Address  `json:"owner_address"`
+	TotalBipStake *big.Int `json:"total_bip_stake"`
+	PubKey        Pubkey   `json:"pub_key"`
+	Commission    uint     `json:"commission"`
+	Stakes        []Stake  `json:"stakes"`
+	Status        byte     `json:"status"`
 }
 
 type Stake struct {
@@ -45,12 +44,12 @@ type Stake struct {
 }
 
 type Coin struct {
-	Name           string     `json:"name"`
-	Symbol         CoinSymbol `json:"symbol"`
-	Volume         *big.Int   `json:"volume"`
-	Crr            uint       `json:"crr"`
-	ReserveBalance *big.Int   `json:"reserve_balance"`
-	MaxSupply      *big.Int   `json:"max_supply"`
+	Name      string     `json:"name"`
+	Symbol    CoinSymbol `json:"symbol"`
+	Volume    *big.Int   `json:"volume"`
+	Crr       uint       `json:"crr"`
+	Reserve   *big.Int   `json:"reserve"`
+	MaxSupply *big.Int   `json:"max_supply"`
 }
 
 type FrozenFund struct {